backend/requirements.txt
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.64
langchain-chroma==0.2.0
chromadb>=0.4.0
tqdm
gdown
pydantic
fastapi
uvicorn[standard]
gunicorn
python-multipart
python-dotenv

# Azure SDK Components (Free Tier Compatible)
azure-data-tables>=12.4.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-monitor-opentelemetry>=1.2.0
opencensus-ext-azure>=1.1.13
applicationinsights>=0.11.10

# Production dependencies
httpx
aiofiles